Renamed stats slice file and added loading to dashboard
