+++ services/frontend-v3/src/redux/slices/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from "redux";
 import settings from "./settingsSlice";
 import errors from "./errorsSlice";
 import user from "./userSlice";
-import stats from "./stats";
+import stats from "./statsSlice";
 import admin from "./adminSlice";
 
 export default combineReducers({
